requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.1
sqlalchemy==2.0.22
pydantic[email]==1.10.11
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.10
